# Conflicts:
#	core/block/backlinks/watcher.go
#	core/block/delete.go
#	core/block/detailservice/set_details.go
#	core/block/editor/basic/details_test.go
#	core/block/editor/file/file_test.go
#	core/block/editor/spaceview.go
#	core/block/object/objectcreator/chat.go
#	core/block/object/objectcreator/creator.go
#	core/block/source/date.go
#	core/block/source/mock_source/mock_Service.go
#	core/block/source/service.go
#	core/domain/genericmap_test.go
#	core/domain/value.go
#	core/object.go
#	core/syncstatus/detailsupdater/updater.go
#	core/syncstatus/detailsupdater/updater_test.go
#	core/syncstatus/syncsubscriptions/syncingobjects.go
#	pkg/lib/localstore/objectstore/fixture.go
#	pkg/lib/localstore/objectstore/spaceindex/fixture.go
#	pkg/lib/localstore/objectstore/spaceindex/objects.go
#	pkg/lib/localstore/objectstore/spaceindex/queries_test.go
#	pkg/lib/localstore/objectstore/spaceindex/store.go
#	pkg/lib/localstore/objectstore/spaceindex/update.go
#	util/pbtypes/anyenc_test.go
#	util/pbtypes/pbtypes.go

+++ core/domain/genericmap_test.go
@@ -436,7 +436,7 @@ func TestGenericMap_ToProto(t *testing.T) {
 			"key-4": pbtypes.Bool(true),
 			"key-5": pbtypes.Bool(false),
 			"key-6": pbtypes.StringList([]string{"1", "2", "3"}),
-			"key-7": pbtypes.FloatList([]float64{1.1, 2.2, 3.3, 4.4}),
+			"key-7": pbtypes.Float64List([]float64{1.1, 2.2, 3.3, 4.4}),
 			"key-8": pbtypes.IntList(1, 2, 3),
 			"key-9": pbtypes.Struct(&types.Struct{
 				Fields: map[string]*types.Value{
